Added helper functions for getting location of storeroom
This patch adds 2 helper functions, one for returning the index of
the storeroom within the list of locations that makes up a map,
the other returns the actual Location object for that storeroom.
In a future patch we can change out instances where we are using
maps.get(p).get(5) to get the storeroom, and stop relying on
hard-coded "5"s around the code.