Fix action command validation
Action command validation wasn't working as intended. This patch
makes it so that the action command is only marked as valid if the
required artefacts, furniture and characters are in the correct
locations out of the player's inventory, storeroom and the current
location. When an artefact is a subject and is marked for
consumption, the artefact is required to be in the player's
inventory. When a furniture is a subject and marked for
consumption, the furniture is required to be in the current
location's furniture list. Every subject marked for production is
required to be in its respective list in the storeroom.

There are 2 exceptions to this:
 * entity: health, either consumed or produced
 * entity: <pathname>, always produced

These exceptions are accounted for in the validation code.